Added stop button, read description
I added a way to stop the songs from playing, but some things to know:
1. Since the program is queuing the songs as they play, it's really
slow, but it should be ok when we make the executable file. Also, when
you press the button to play, the song takes about 3 seconds to play.
This intentional, it's just the way it has to be.
2. I changed the on-tick function, so take a look at it if you're
working on the metronome.